Move NTP show command to separate file
